Fix structure and spelling in all files
1. Removed parentheses around conditionals in while and if statements,
since it's more julian.
2. Made spelling more julian.
3. Changed tab size from 4 to 2, since most packages use this as
standard.
4. Respelled `schduler` to `scheduler` in exports.
5. Removed `line` from exports.
6. Moved Game-utilities from `GameZero.jl` to a separate file `Game.jl`,
and thus `GameZero.jl` only contains exports, imports and includes.
7. Restructured `GameZero.jl` to make it clear what it depend upon and
exports.
8. Moved Geom-utilities from `screen.jl` to a separate file `Geom.jl`.
9. Renamed `screen.jl` to `Screen.jl` since it represents a struct.
10. Made a folder for keyboard-utilities.
11. Moved every keyboard module to their own module-file.
12. Renamed `keyboard.jl` to `Keyboard.jl`, since it only contained
functionality for the Keyboard struct.
13. Made a folder for timer-utilities.
14. Created a file in `timer/` for WallTimer-utilities, and removed the
same content in `timer.jl`.
15. Created a file in `timer/` for Scheduled structs and utilities, and
removed the file `timer.jl`.
Creating modules to be separate files
